chore: refactor profile page
